fixing directory structure
