+++ .env.example
@@ -1,0 +1,4 @@
+CODECHEF_CLIENT_ID=
+CODECHEF_CLIENT_SECRET=
+DISCORD_BOT_TOKEN=
+CODEFORCES_API_KEY=
