+++ .gitignore
@@ -1,0 +1,4 @@
+.env
+.vscode
+parakh/
+__pycache__/
